refactored event creation into a form to reusable for event editing. added event editing component.
